turn UnpublishedTweetsProvier to Resolver, that allows opening for testing
